Move cdn models to jvm library
