+++ cdn/src/main/java/com/urlaunched/android/cdn/domain/repository/OriginalFilePreviewLinkRepository.kt
@@ -1,7 +1,7 @@
 package com.urlaunched.android.cdn.domain.repository
 
-import com.urlaunched.android.cdn.models.domain.downloadstate.DownloadState
-import com.urlaunched.android.cdn.models.domain.links.MediaLink
+import com.urlaunched.android.cdnmodels.domain.downloadstate.DownloadState
+import com.urlaunched.android.cdnmodels.domain.links.MediaLink
 import com.urlaunched.android.common.response.Response
 import kotlinx.coroutines.flow.Flow
