+++ cdn/src/main/java/com/urlaunched/android/cdn/domain/usecases/DownloadFileUseCase.kt
@@ -1,7 +1,7 @@
 package com.urlaunched.android.cdn.domain.usecases
 
 import com.urlaunched.android.cdn.domain.repository.OriginalFilePreviewLinkRepository
-import com.urlaunched.android.cdn.models.domain.links.MediaLink
+import com.urlaunched.android.cdnmodels.domain.links.MediaLink
 
 class DownloadFileUseCase(private val repository: OriginalFilePreviewLinkRepository) {
     suspend operator fun invoke(mediaLink: MediaLink, path: String) = repository.downloadFile(mediaLink, path)
