+++ cdn/src/main/java/com/urlaunched/android/cdn/domain/usecases/GetOriginalFileLinkUseCase.kt
@@ -1,7 +1,7 @@
 package com.urlaunched.android.cdn.domain.usecases
 
 import com.urlaunched.android.cdn.domain.repository.OriginalFilePreviewLinkRepository
-import com.urlaunched.android.cdn.models.domain.links.MediaLink
+import com.urlaunched.android.cdnmodels.domain.links.MediaLink
 
 class GetOriginalFileLinkUseCase(private val repository: OriginalFilePreviewLinkRepository) {
     suspend operator fun invoke(mediaLink: MediaLink.Private) = repository.getPrivateFileLink(mediaLink)
