+++ cdnmodels/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=cdnmodels
+POM_NAME=Common cdn models
+POM_PACKAGING=aar
+VERSION_NAME=LOCAL
