+++ cdnmodels/src/main/java/com/urlaunched/android/cdnmodels/domain/links/MediaLink.kt
@@ -1,4 +1,4 @@
-package com.urlaunched.android.cdn.models.domain.links
+package com.urlaunched.android.cdnmodels.domain.links
 
 import kotlinx.serialization.Serializable
 
@@ -6,19 +6,19 @@ import kotlinx.serialization.Serializable
 sealed class MediaLink {
     abstract val id: Int
     abstract val link: String
-    abstract val sizeKb: Int
+    abstract val sizeKb: Int?
 
     @Serializable
     data class Public(
         override val id: Int,
         override val link: String,
-        override val sizeKb: Int
+        override val sizeKb: Int?
     ) : MediaLink()
 
     @Serializable
     data class Private(
         override val id: Int,
         override val link: String,
-        override val sizeKb: Int
+        override val sizeKb: Int?
     ) : MediaLink()
 }
