+++ cdnmodels/src/main/java/com/urlaunched/android/cdnmodels/remote/MediaRemoteModel.kt
@@ -1,6 +1,6 @@
-package com.urlaunched.android.cdn.models.remote
+package com.urlaunched.android.cdnmodels.remote
 
-import com.urlaunched.android.cdn.models.domain.media.MediaDomainModel
+import com.urlaunched.android.cdnmodels.domain.media.MediaDomainModel
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -9,11 +9,11 @@ data class MediaRemoteModel(
     @SerialName("id")
     val id: Int,
     @SerialName("media_type")
-    val mediaType: String,
+    val mediaType: String?,
     @SerialName("media_url")
     val mediaUrl: String,
     @SerialName("size_kb")
-    val sizeKb: Int
+    val sizeKb: Int?
 )
 
 fun MediaRemoteModel.toCDNMedia() = MediaDomainModel(
