+++ settings.gradle
@@ -30,3 +30,4 @@ include ':network'
 include ':cdn'
 include ':tempattachment'
 include ':snapshottesting'
+include ':cdnmodels'
